Add ability to 'debug' hooks with `objection explore -d`
This commit also contains a small refactor of the template
cleanup processor, moving it to the FridaRunner class at
`objection.utils.frida_transport.FridaRunner#_hook_processor`.
This is also where the check for debugging hooks is made.

A new dependency is also introduced (jsbeautifier) that will
pretty format hooks before dumping them to the screen & log.