Refactor the IPA patcher.
The process of patching an iOS IPA involves quite a number of
steps. This easily results in along blocks of procedural code
getting written. This commit refactors the peices needed to
patch an IPA into a number of classes that have components that
could be reused in the Android patcher.