Properly pull overload from method, set return properly
This works on the splitClassMethod to pull out an overload, if one was
supplied by the user. If so, a new overload is returned by the function
for use in the set-return-value function (uses the content in an eval()
loop - wasn't sure how else to dynamically parse that content).

TODO added for considering implementing overload in watch-method, etc.

Works on sensepost/objection#206