Refactor Android heap interactions.
The frida-java-bridge updated and renamed some internals that we
were mistakenly relying on. This commit refactors the Andriod
heap interactions to make it compatible with the latest version of
frida-java-bridge.

Most importantly, object references are now made using hashcodes[1]
instead of the internal $handle reference (now renamed to $h).

Lastly, the `--fresh` flag was also removed as it was confusing. Every
call to `android heap search instances` will make fresh calls.

[1] https://developer.android.com/reference/java/lang/Object#hashCode()