format Apktool decode command line parameters properly
Two command line parameters were added in a buggy way that resulted in a
broken command line if those parameters were not needed, as instead of
skipping the argument, the code added an empty string as an argument,
which (at least certain recent versions of) Apktool refused to accept.

This commit changes the behavior to be in line with other parts of the
code (e.g., Apktool build).